updates the WPF to include data grid and reference points

+++ DistanceWPF/Model/ReferencePoint.cs
@@ -8,7 +8,7 @@ namespace Distance.Model
 {
     class ReferencePoint : DistancePoint
     {
-        public ReferencePoint(int x, int y)
+        public ReferencePoint(double x, double y)
             : base(x, y)
         { }
     }
